Timeslot shift entity adjustments
