+++ LBH.AdultSocialCare.Api/V1/Constants.cs
@@ -1,4 +1,4 @@
-namespace LBH.AdultSocialCare.Api.V1.Controllers
+namespace LBH.AdultSocialCare.Api.V1
 {
     public static class Constants
     {
