+++ LBH.AdultSocialCare.Api/V1/CorrelationMiddleware.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 
-namespace LBH.AdultSocialCare.Api.V1.Controllers
+namespace LBH.AdultSocialCare.Api.V1
 {
     public class CorrelationMiddleware
     {
