+++ LBH.AdultSocialCare.Api/V1/Factories/HomeCarePackageSlotsFactory.cs
@@ -3,6 +3,8 @@ using LBH.AdultSocialCare.Api.V1.Boundary.Response;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
 using System.Linq;
+using LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare;
+using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.Factories
 {
