+++ LBH.AdultSocialCare.Api/V1/Gateways/Interfaces/IHomeCareServiceTypeGateway.cs
@@ -2,6 +2,7 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.Gateways.Interfaces
 {
