+++ LBH.AdultSocialCare.Api/V1/Infrastructure/Entities/HomeCare/HomeCarePackage.cs
@@ -1,20 +1,29 @@
 using System;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
-namespace LBH.AdultSocialCare.Api.V1.Boundary.Request
+namespace LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare
 {
 
-    public class HomeCarePackageRequest
+    public class HomeCarePackage : BaseEntity
     {
 
         /// <summary>
         /// Gets or sets the Id
         /// </summary>
+        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public Guid Id { get; set; }
 
         /// <summary>
         /// Gets or sets the Client Id
         /// </summary>
         public Guid ClientId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Clients
+        /// </summary>
+        [ForeignKey(nameof(ClientId))]
+        public Clients Clients { get; set; }
 
         /// <summary>
         /// Gets or sets the Start Date
@@ -60,6 +69,11 @@ namespace LBH.AdultSocialCare.Api.V1.Boundary.Request
         /// Gets or sets the Status Id
         /// </summary>
         public Guid StatusId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Status Object
+        /// </summary>
+        public Status Status { get; set; }
 
     }
