+++ LBH.AdultSocialCare.Api/V1/Infrastructure/Entities/HomeCare/HomeCarePackageSlots.cs
@@ -1,26 +1,35 @@
 using System;
-using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
-namespace LBH.AdultSocialCare.Api.V1.Boundary.Request
+namespace LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare
 {
 
-    public class HomeCarePackageSlotsRequestList
+    public class HomeCarePackageSlots
     {
 
         /// <summary>
         /// Gets or sets the Id
         /// </summary>
+        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public Guid Id { get; set; }
 
         /// <summary>
         /// Gets or sets the Home Care Package Id
         /// </summary>
+        [ForeignKey(nameof(HomeCarePackageId))]
         public Guid HomeCarePackageId { get; set; }
 
         /// <summary>
         /// Gets or sets the Service Id
         /// </summary>
         public Guid ServiceId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Services
+        /// </summary>
+        [ForeignKey(nameof(ServiceId))]
+        public HomeCareServiceType Services { get; set; }
 
         /// <summary>
         /// Gets or sets the Primary Carer
@@ -43,14 +52,26 @@ namespace LBH.AdultSocialCare.Api.V1.Boundary.Request
         public string WhatShouldBeDone { get; set; }
 
         /// <summary>
-        /// Gets or sets the Home Care Package Slot Response
+        /// Gets or sets the Time Slot Type Id
         /// </summary>
-        public List<HomeCarePackageSlotRequest> HomeCarePackageSlotRequest { get; set; }
+        public Guid TimeSlotTypeId { get; set; }
 
-    }
+        /// <summary>
+        /// Gets or sets the Time Slot Types
+        /// </summary>
+        [ForeignKey(nameof(TimeSlotTypeId))]
+        public TimeSlotType TimeSlotTypes { get; set; }
 
-    public class HomeCarePackageSlotRequest
-    {
+        /// <summary>
+        /// Gets or sets the Time Slot Shift Id
+        /// </summary>
+        public int TimeSlotShiftId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the Time Slot Shift
+        /// </summary>
+        [ForeignKey(nameof(TimeSlotShiftId))]
+        public TimeSlotShifts TimeSlotShift { get; set; }
 
         /// <summary>
         /// Gets or sets the InMinutes
@@ -66,16 +87,6 @@ namespace LBH.AdultSocialCare.Api.V1.Boundary.Request
         /// Gets or sets the Time
         /// </summary>
         public int Time { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Time Slot Type Id
-        /// </summary>
-        public Guid TimeSlotTypeId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Time Slot Shift Id
-        /// </summary>
-        public int TimeSlotShiftId { get; set; }
 
     }
