+++ LBH.AdultSocialCare.Api/V1/Infrastructure/Entities/HomeCare/TimeSlotShifts.cs
@@ -1,8 +1,6 @@
-using System;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
-namespace LBH.AdultSocialCare.Api.V1.Infrastructure.Entities
+namespace LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare
 {
 
     public class TimeSlotShifts : BaseEntity
