+++ LBH.AdultSocialCare.Api/V1/Infrastructure/Entities/TermTimeConsiderationOption.cs
@@ -3,11 +3,17 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace LBH.AdultSocialCare.Api.V1.Infrastructure.Entities
 {
+
     public class TermTimeConsiderationOption
     {
+
         [Key]
         [Column("OptionId")]
         public int OptionId { get; set; }
-        public string OptionName { get; set; } // N/A, Term Time, Holiday
+
+        // N/A, Term Time, Holiday
+        public string OptionName { get; set; }
+
     }
+
 }
