+++ LBH.AdultSocialCare.Api/V1/UseCase/GetAllTimeSlotTypesUseCase.cs
@@ -3,6 +3,7 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase
 {
