+++ LBH.AdultSocialCare.Api/V1/UseCase/Interfaces/IGetAllTimeSlotShiftsUseCase.cs
@@ -1,7 +1,7 @@
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-
+using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
 {
