+++ LBH.AdultSocialCare.Api/V1/UseCase/Interfaces/IGetAllTimeSlotTypesUseCase.cs
@@ -1,6 +1,7 @@
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
 {
